+++ DarkCodex/changelog.md
@@ -1,4 +1,9 @@
 # Changelog
+- ~~fixed blade rush this time?~~
+- fixed duplicate undead immunity that blocked dirge bard archetype and undead bloodline
+- added Mythic.createSwiftHuntersBond
+- added Ranger.createImprovedHuntersBond
+- added metamagic to preferred spell 
 
 ## [1.0.8]
 - added General.patchVarious
@@ -9,7 +14,7 @@
 - General various: extend protection from X to 10 minutes
 - fixed gather power blocked during polymorph, as well as kitsune human form being unable to use kinetic blade
 - fixed feral combat applying to all weapon checks, instead of only unarmed checks
-- fixed blade rush attacking at the start point, instead of the end point
+- ~~fixed blade rush attacking at the start point, instead of the end point~~
 - fixed preferred spell
 - added debug options to enable inventory while polymorphed
 - fixed resourceful caster now refunds converted spells